Feat: menu restaurant

+++ database/migrations/2022_04_15_185617_create_menus_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->binary('image');
-            $table->tinyInteger('price');
+            $table->integer('price');
             $table->string('category');
             $table->timestamps();
